Outline of the firmware is working
..complete with a key matrix, etc.

Have a problem whereby the keyboard stops (panics?) if
I press numlock or capslock on the laptop keyboard.  I
assume it's something related to the laptop sending LED
status that I'm not handling - but not sure.

If I don't press numlock or caps lock on the laptop,
it seems to work just fine (but only 3 letters configured
so far).