feat: Debounce is more responsive and more aggressive at removing double presses.
Implement a different debounce strategy as I subjectively
felt the previous one was giving lots of false double-presses.

The previous strategy used the concept of a floating value which
would be dragged upward or downward based on the input signal
and only report a change when the floating value saturated.

I found it difficult to strike a balance between responsiveness
(time to register first press) and elimnination of false presses.
Increasing the gap between the min and max values would address
double presses but at the cost of increasing latency.

New approach is simpler.  We report changes _immediately_ but
then have a cooldown period during which any further change will
be ignored.

This different strategy allows me to have quite a long cooldown
(50ms) which eliminates double presses, but does not introduce
latency. For comparison the previous strategy had a gap between
min and max of approx 10ms.

Brief accidental presses are more likely to register now, so will
keep an eye out for that.  It might be that _both_ strategies together
is what would work best..